Doco and embedding logic in components (#29)
* Switch to a package-info for prism generation

It is cleaner than a class

* Add some more in-code documentation

Including adding the `NullMarked` annotations... which I thought were there but apparently not!

* Embed unwrap logic in component

This starts work on #25, although it's not finished. I'm not sure if I am committed to this, and it may need to take in another variable for e.g. `WriteCharSequence`...

* Fix javadoc build, then the actual build

Also add a helper shell script that runs everything the same way that the build does, fix the cache in said build.

* Address PR comments

* More PR comments

* Final PR comments